src/main/webapp/app/entities/tren/route/tren-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { ITren, Tren } from '../tren.model';
import { TrenService } from '../service/tren.service';

import { TrenRoutingResolveService } from './tren-routing-resolve.service';

describe('Tren routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TrenRoutingResolveService;
  let service: TrenService;
  let resultTren: ITren | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(TrenRoutingResolveService);
    service = TestBed.inject(TrenService);
    resultTren = undefined;
  });

  describe('resolve', () => {
    it('should return ITren returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTren = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTren).toEqual({ id: 123 });
    });

    it('should return new ITren if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTren = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTren).toEqual(new Tren());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Tren })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTren = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTren).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
